requirements.txt
# -*- coding: UTF-8 -*-
#requirement.txt
Django==1.7.5
bootstrap-admin
django-pagedown==0.1.0
django-markdown-deux==1.0.5
djangorestframework==3.0.5

#Install production dependencies only.
#-r requirements/prod.txt